Fix home-manager issue with google chrome
I got the following error:

The option `programs.google-chrome.extensions' can no longer be used
since it's been removed. The `extensions` option does not work on
Google Chrome anymore.

It is caused by the following commit:

https://github.com/nix-community/home-manager/commit/6245dd11cf59698aacdf8b153e3e7070abbef36f

So I decided not to use the home-manager module for google-chrome and
just add the package. I need to investigate this further to reenable
this module.